+++ packages/inter-protocol/src/proposals/addAssetToVault.js
@@ -1,11 +1,12 @@
 // @jessie-check
 
 import { AmountMath, AssetKind } from '@agoric/ertp';
-import { makeRatio } from '@agoric/zoe/src/contractSupport/index.js';
 import { deeplyFulfilledObject } from '@agoric/internal';
 import { Stable } from '@agoric/vats/src/tokens.js';
-import { E } from '@endo/far';
+import { makeRatio } from '@agoric/zoe/src/contractSupport/index.js';
 import { parseRatio } from '@agoric/zoe/src/contractSupport/ratio.js';
+import { E } from '@endo/far';
+import { instanceNameFor } from './price-feed-proposal.js';
 import { reserveThenGetNames } from './utils.js';
 
 export * from './startPSM.js';
@@ -211,7 +212,7 @@ export const registerScaledPriceAuthority = async (
  * @param {object} config.options
  * @param {InterchainAssetOptions} config.options.interchainAssetOptions
  * @param {bigint | number | string} config.options.debtLimitValue
- * @param {bigint} config.options.interestRateValue
+ * @param {bigint} config.options.stabilityFeeValue
  */
 export const addAssetToVault = async (
   {
@@ -219,6 +220,7 @@ export const addAssetToVault = async (
     brand: {
       consume: { [Stable.symbol]: stableP },
     },
+    instance: { consume: consumeInstance },
   },
   {
     options: {
@@ -226,7 +228,7 @@ export const addAssetToVault = async (
       debtLimitValue = 1_000n * 1_000_000n,
       // Default to a safe value. Production will likely set this through governance.
       // Allow setting through bootstrap to simplify testing.
-      interestRateValue = 1n,
+      stabilityFeeValue = 1n,
       interchainAssetOptions,
     },
   },
@@ -239,11 +241,16 @@ export const addAssetToVault = async (
     [keyword],
   );
 
+  const oracleInstanceName = instanceNameFor(oracleBrand, 'USD');
+  // don't add the collateral offering to vaultFactory until its price feed is available
+  // eslint-disable-next-line no-restricted-syntax -- allow this computed property
+  await consumeInstance[oracleInstanceName];
+
   const stable = await stableP;
   const vaultFactoryCreator = E.get(vaultFactoryKit).creatorFacet;
   await E(vaultFactoryCreator).addVaultType(interchainIssuer, oracleBrand, {
     debtLimit: AmountMath.make(stable, BigInt(debtLimitValue)),
-    interestRate: makeRatio(interestRateValue, stable),
+    stabilityFee: makeRatio(stabilityFeeValue, stable),
     // The rest of these we use safe defaults.
     // In production they will be governed by the Econ Committee.
     // Product deployments are also expected to have a low debtLimitValue at the outset,
@@ -261,7 +268,7 @@ export const getManifestForAddAssetToVault = (
   { restoreRef },
   {
     debtLimitValue,
-    interestRateValue,
+    stabilityFeeValue,
     interchainAssetOptions,
     scaledPriceAuthorityRef,
   },
@@ -321,6 +328,11 @@ export const getManifestForAddAssetToVault = (
         brand: {
           consume: { [Stable.symbol]: true },
         },
+        instance: {
+          // allow any instance because the AGORIC_INSTANCE_NAME of
+          // priceFeedOptions cannot be known statically.
+          consume: true,
+        },
       },
     },
     installations: {
@@ -329,7 +341,7 @@ export const getManifestForAddAssetToVault = (
     options: {
       debtLimitValue,
       interchainAssetOptions,
-      interestRateValue,
+      stabilityFeeValue,
     },
   };
 };