+++ packages/inter-protocol/src/vaultFactory/params.js
@@ -21,7 +21,7 @@ export const DEBT_LIMIT_KEY = 'DebtLimit';
 export const LIQUIDATION_MARGIN_KEY = 'LiquidationMargin';
 export const LIQUIDATION_PADDING_KEY = 'LiquidationPadding';
 export const LIQUIDATION_PENALTY_KEY = 'LiquidationPenalty';
-export const INTEREST_RATE_KEY = 'InterestRate';
+export const STABILITY_FEE_KEY = 'StabilityFee';
 export const MINT_FEE_KEY = 'MintFee';
 export const MIN_INITIAL_DEBT_KEY = 'MinInitialDebt';
 export const SHORTFALL_INVITATION_KEY = 'ShortfallInvitation';
@@ -89,7 +89,7 @@ export const makeVaultParamManager = (
   publisherKit,
   {
     debtLimit,
-    interestRate,
+    stabilityFee,
     liquidationMargin,
     liquidationPadding = zeroRatio(liquidationMargin),
     liquidationPenalty,
@@ -98,7 +98,7 @@ export const makeVaultParamManager = (
 ) =>
   makeParamManagerSync(publisherKit, {
     [DEBT_LIMIT_KEY]: [ParamTypes.AMOUNT, debtLimit],
-    [INTEREST_RATE_KEY]: [ParamTypes.RATIO, interestRate],
+    [STABILITY_FEE_KEY]: [ParamTypes.RATIO, stabilityFee],
     [LIQUIDATION_PADDING_KEY]: [ParamTypes.RATIO, liquidationPadding],
     [LIQUIDATION_MARGIN_KEY]: [ParamTypes.RATIO, liquidationMargin],
     [LIQUIDATION_PENALTY_KEY]: [ParamTypes.RATIO, liquidationPenalty],
@@ -110,7 +110,7 @@ export const vaultParamPattern = M.splitRecord(
   {
     liquidationMargin: ratioPattern,
     liquidationPenalty: ratioPattern,
-    interestRate: ratioPattern,
+    stabilityFee: ratioPattern,
     mintFee: ratioPattern,
     debtLimit: amountPattern,
   },