+++ packages/inter-protocol/test/vaultFactory/test-collatRatio.js
@@ -21,7 +21,7 @@ test('normalizedCollRatio grows with coll margin', t => {
   /** @type {PriceQuote} */
   // @ts-expect-error fake for tests.
   const quote = makeFakeQuote(minted.make(5n), collateral.make(20n));
-  const compoundedInterest = makeRatioFromAmounts(
+  const compoundedStabilityFee = makeRatioFromAmounts(
     minted.make(1n),
     collateral.make(1n),
   );
@@ -36,13 +36,13 @@ test('normalizedCollRatio grows with coll margin', t => {
 
   const lowRate = normalizedCollRatio(
     quote,
-    compoundedInterest,
+    compoundedStabilityFee,
     lowCollateralizationMargin,
   );
   t.deepEqual(lowRate, (105 * 5) / (20 * 100));
   const highRate = normalizedCollRatio(
     quote,
-    compoundedInterest,
+    compoundedStabilityFee,
     highCollateralizationMargin,
   );
   t.deepEqual(highRate, (150 * 5) / (20 * 100));
@@ -78,15 +78,19 @@ test('CollRatio grows with price', t => {
   // @ts-expect-error fake for tests.
   const highQuote = makeFakeQuote(minted.make(15n), collateral.make(20n));
 
-  const compoundedInterest = makeRatioFromAmounts(
+  const compoundedStabilityFee = makeRatioFromAmounts(
     minted.make(12n),
     collateral.make(10n),
   );
   const margin = makeRatioFromAmounts(minted.make(105n), collateral.make(100n));
 
-  const lowRate = normalizedCollRatio(lowQuote, compoundedInterest, margin);
+  const lowRate = normalizedCollRatio(lowQuote, compoundedStabilityFee, margin);
   t.deepEqual(lowRate, (105 * 5 * 12) / (20 * 100 * 10));
-  const highRate = normalizedCollRatio(highQuote, compoundedInterest, margin);
+  const highRate = normalizedCollRatio(
+    highQuote,
+    compoundedStabilityFee,
+    margin,
+  );
   t.deepEqual(highRate, (105 * 15 * 12) / (20 * 100 * 10));
   t.true(highRate > lowRate);
 });