+++ packages/inter-protocol/test/vaultFactory/test-vault-interest.js
@@ -23,7 +23,7 @@ const trace = makeTracer('TestVaultInterest', false);
  * @property {IssuerKit} collateralKit
  * @property {Vault} vault
  * @property {Function} advanceRecordingPeriod
- * @property {Function} setInterestRate
+ * @property {Function} setStabilityFee
  */
 let testJig;
 /** @param {TestContext} jig */
@@ -137,7 +137,7 @@ test('charges', async t => {
     AmountMath.make(stableBrand, startingDebt - normalizedPaybackValue),
   );
 
-  testJig.setInterestRate(25n);
+  testJig.setStabilityFee(25n);
 
   for (const [i, charge] of [22n, 27n, 34n].entries()) {
     // XXX https://github.com/Agoric/agoric-sdk/issues/5527