+++ packages/ui-components/src/index.js
@@ -1,8 +1,0 @@
-/* eslint-disable import/no-unresolved */
-export { default as makeNatAmountInput } from './components/NatAmountInput.js';
-// eslint-disable-next-line import/no-unresolved
-export {
-  default as makeNatPurseSelector,
-  filterPurses,
-} from './components/NatPurseSelector.js';
-export * from './display/index.js';