+++ packages/vats/test/bootstrapTests/test-liquidation-1.js
@@ -1,6 +1,6 @@
 // @ts-check
 /**
- * @file Bootstrap test integration vaults with smart-wallet
+ * @file Bootstrap test of liquidation across multiple collaterals
  */
 import { test as anyTest } from '@agoric/zoe/tools/prepare-test-env-ava.js';
 
@@ -9,6 +9,7 @@ import {
   makeParseAmount,
   Offers,
 } from '@agoric/inter-protocol/src/clientSupport.js';
+import process from 'process';
 import {
   likePayouts,
   makeLiquidationTestContext,
@@ -19,9 +20,6 @@ import {
  * @type {import('ava').TestFn<Awaited<ReturnType<typeof makeLiquidationTestContext>>>}
  */
 const test = anyTest;
-
-// presently all these tests use one collateral manager
-const collateralBrandKey = 'ATOM';
 
 //#region Product spec
 const setup = /** @type {const} */ ({
@@ -92,6 +90,7 @@ const outcome = /** @type {const} */ ({
   reserve: {
     allocations: {
       ATOM: 0.309852,
+      STARS: 0.309852,
     },
     shortfall: 0,
   },
@@ -129,24 +128,38 @@ test.after.always(t => {
 });
 
 // Reference: Flow 1 from https://github.com/Agoric/agoric-sdk/issues/7123
-test.serial('scenario: Flow 1', async t => {
+const checkFlow1 = async (
+  t,
+  { collateralBrandKey, managerIndex },
+  _expected,
+) => {
+  // fail if there are any unhandled rejections
+  process.on('unhandledRejection', error => {
+    t.fail(/** @type {Error} */ (error).message);
+  });
+
   const {
     advanceTimeBy,
     advanceTimeTo,
     agoricNamesRemotes,
     check,
     setupStartingState,
-    priceFeedDriver,
+    priceFeedDrivers,
     readLatest,
     walletFactoryDriver,
   } = t.context;
 
-  await setupStartingState();
+  const metricsPath = `published.vaultFactory.managers.manager${managerIndex}.metrics`;
+
+  await setupStartingState({
+    collateralBrandKey,
+    managerIndex,
+  });
 
   const minter = await walletFactoryDriver.provideSmartWallet('agoric1minter');
 
   for (let i = 0; i < setup.vaults.length; i += 1) {
-    const offerId = `open-vault${i}`;
+    const offerId = `open-${collateralBrandKey}-vault${i}`;
     await minter.executeOfferMaker(Offers.vaults.OpenVault, {
       offerId,
       collateralBrandKey,
@@ -161,7 +174,7 @@ test.serial('scenario: Flow 1', async t => {
 
   // Verify starting balances
   for (let i = 0; i < setup.vaults.length; i += 1) {
-    check.vaultNotification(i, {
+    check.vaultNotification(managerIndex, i, {
       debtSnapshot: { debt: { value: scale6(setup.vaults[i].debt) } },
       locked: { value: scale6(setup.vaults[i].atom) },
       vaultState: 'active',
@@ -180,18 +193,18 @@ test.serial('scenario: Flow 1', async t => {
         agoricNamesRemotes.instance['psm-IST-USDC_axl'],
         agoricNamesRemotes.brand,
         {
-          offerId: 'print-ist',
+          offerId: `print-${collateralBrandKey}-ist`,
           wantMinted: 1_000,
           pair: ['IST', 'USDC_axl'],
         },
       ),
     );
 
-    const maxBuy = '10000ATOM';
+    const maxBuy = `10000${collateralBrandKey}`;
 
     // bids are long-lasting offers so we can't wait here for completion
     await buyer.sendOfferMaker(Offers.auction.Bid, {
-      offerId: 'bid1',
+      offerId: `${collateralBrandKey}-bid1`,
       ...setup.bids[0],
       maxBuy,
       parseAmount,
@@ -199,33 +212,33 @@ test.serial('scenario: Flow 1', async t => {
 
     t.like(readLatest('published.wallet.agoric1buyer'), {
       status: {
-        id: 'bid1',
+        id: `${collateralBrandKey}-bid1`,
         result: 'Your bid has been accepted',
         payouts: undefined,
       },
     });
     await buyer.sendOfferMaker(Offers.auction.Bid, {
-      offerId: 'bid2',
+      offerId: `${collateralBrandKey}-bid2`,
       ...setup.bids[1],
       maxBuy,
       parseAmount,
     });
     t.like(readLatest('published.wallet.agoric1buyer'), {
       status: {
-        id: 'bid2',
+        id: `${collateralBrandKey}-bid2`,
         result: 'Your bid has been accepted',
         payouts: undefined,
       },
     });
     await buyer.sendOfferMaker(Offers.auction.Bid, {
-      offerId: 'bid3',
+      offerId: `${collateralBrandKey}-bid3`,
       ...setup.bids[2],
       maxBuy,
       parseAmount,
     });
     t.like(readLatest('published.wallet.agoric1buyer'), {
       status: {
-        id: 'bid3',
+        id: `${collateralBrandKey}-bid3`,
         result: 'Your bid has been accepted',
         payouts: undefined,
       },
@@ -237,22 +250,21 @@ test.serial('scenario: Flow 1', async t => {
     //  Change price to trigger liquidation
     // ---------------
 
-    await priceFeedDriver.setPrice(9.99);
+    await priceFeedDrivers[collateralBrandKey].setPrice(9.99);
 
     // check nothing liquidating yet
     /** @type {import('@agoric/inter-protocol/src/auction/scheduler.js').ScheduleNotification} */
     const liveSchedule = readLatest('published.auction.schedule');
     t.is(liveSchedule.activeStartTime, null);
-    t.like(readLatest('published.vaultFactory.managers.manager0.metrics'), {
+    t.like(readLatest(metricsPath), {
       numActiveVaults: setup.vaults.length,
       numLiquidatingVaults: 0,
-      lockedQuote: null,
     });
 
     // advance time to start an auction
-    console.log('step 0 of 10');
+    console.log(collateralBrandKey, 'step 1 of 10');
     await advanceTimeTo(NonNullish(liveSchedule.nextDescendingStepTime));
-    t.like(readLatest('published.vaultFactory.managers.manager0.metrics'), {
+    t.like(readLatest(metricsPath), {
       numActiveVaults: 0,
       numLiquidatingVaults: setup.vaults.length,
       liquidatingCollateral: {
@@ -262,24 +274,24 @@ test.serial('scenario: Flow 1', async t => {
       lockedQuote: null,
     });
 
-    console.log('step 1 of 10');
+    console.log(collateralBrandKey, 'step 2 of 10');
     await advanceTimeBy(3, 'minutes');
-    t.like(readLatest('published.auction.book0'), {
+    t.like(readLatest(`published.auction.book${managerIndex}`), {
       collateralAvailable: { value: scale6(setup.auction.start.collateral) },
       startCollateral: { value: scale6(setup.auction.start.collateral) },
       startProceedsGoal: { value: scale6(setup.auction.start.debt) },
     });
 
-    console.log('step 2 of 10');
+    console.log(collateralBrandKey, 'step 3 of 10');
     await advanceTimeBy(3, 'minutes');
 
-    console.log('step 3 of 10');
+    console.log(collateralBrandKey, 'step 4 of 10');
     await advanceTimeBy(3, 'minutes');
     // XXX updates for bid1 and bid2 are appended in the same turn so readLatest gives bid2
     // NB: console output shows 8897786n payout which matches spec 8.897ATOM
     // t.like(readLatest('published.wallet.agoric1buyer'), {
     //   status: {
-    //     id: 'bid1',
+    //     id: `${collateralBrandKey}-bid1`,
     //     payouts: {
     //       Bid: { value: 0n },
     //       Collateral: { value: scale6(outcome.bids[0].payouts.Collateral) },
@@ -289,30 +301,30 @@ test.serial('scenario: Flow 1', async t => {
 
     t.like(readLatest('published.wallet.agoric1buyer'), {
       status: {
-        id: 'bid2',
+        id: `${collateralBrandKey}-bid2`,
         payouts: likePayouts(outcome.bids[1].payouts),
       },
     });
 
-    console.log('step 4 of 10');
+    console.log(collateralBrandKey, 'step 5 of 10');
     await advanceTimeBy(3, 'minutes');
 
-    console.log('step 5 of 10');
+    console.log(collateralBrandKey, 'step 6 of 10');
     await advanceTimeBy(3, 'minutes');
-    t.like(readLatest('published.auction.book0'), {
+    t.like(readLatest(`published.auction.book${managerIndex}`), {
       collateralAvailable: { value: 9659301n },
     });
 
-    console.log('step 6 of 10');
+    console.log(collateralBrandKey, 'step 7 of 10');
     await advanceTimeBy(3, 'minutes');
 
-    console.log('step 7 of 10');
+    console.log(collateralBrandKey, 'step 8 of 10');
     await advanceTimeBy(3, 'minutes');
 
-    console.log('step 8 of 10');
+    console.log(collateralBrandKey, 'step 9 of 10');
     await advanceTimeBy(3, 'minutes');
     // Not part of product spec
-    t.like(readLatest('published.vaultFactory.managers.manager0.metrics'), {
+    t.like(readLatest(metricsPath), {
       numActiveVaults: 0,
       numLiquidationsCompleted: setup.vaults.length,
       numLiquidatingVaults: 0,
@@ -325,36 +337,37 @@ test.serial('scenario: Flow 1', async t => {
       totalShortfallReceived: { value: 0n },
     });
 
-    console.log('step 9 of 10');
-    await advanceTimeBy(3, 'minutes');
-
-    console.log('step 10 of 10');
-    await advanceTimeBy(3, 'minutes');
-
-    console.log('step 11 of 10');
-    await advanceTimeBy(3, 'minutes');
+    console.log(collateralBrandKey, 'step 10 of 10');
+    // continuing after now would start a new auction
+    {
+      /** @type {Record<string, import('@agoric/time/src/types.js').TimestampRecord>} */
+      const { nextDescendingStepTime, nextStartTime } = readLatest(
+        'published.auction.schedule',
+      );
+      t.is(nextDescendingStepTime.absValue, nextStartTime.absValue);
+    }
 
     // bid3 still live because it's not fully satisfied
     const { liveOffers } = readLatest('published.wallet.agoric1buyer.current');
-    t.is(liveOffers[0][1].id, 'bid3');
+    t.is(liveOffers[0][1].id, `${collateralBrandKey}-bid3`);
     // exit to get payouts
-    await buyer.tryExitOffer('bid3');
+    await buyer.tryExitOffer(`${collateralBrandKey}-bid3`);
     t.like(readLatest('published.wallet.agoric1buyer'), {
       status: {
-        id: 'bid3',
+        id: `${collateralBrandKey}-bid3`,
         payouts: likePayouts(outcome.bids[2].payouts),
       },
     });
 
     // TODO express spec up top in a way it can be passed in here
-    check.vaultNotification(0, {
+    check.vaultNotification(managerIndex, 0, {
       debt: undefined,
       vaultState: 'liquidated',
       locked: {
         value: scale6(outcome.vaultsActual[0].locked),
       },
     });
-    check.vaultNotification(1, {
+    check.vaultNotification(managerIndex, 1, {
       debt: undefined,
       vaultState: 'liquidated',
       locked: {
@@ -366,10 +379,58 @@ test.serial('scenario: Flow 1', async t => {
   // check reserve balances
   t.like(readLatest('published.reserve.metrics'), {
     allocations: {
-      ATOM: { value: scale6(outcome.reserve.allocations.ATOM) },
-      // not part of product spec
-      Fee: { value: scale6(1.54) },
+      [collateralBrandKey]: {
+        value: scale6(outcome.reserve.allocations[collateralBrandKey]),
+      },
     },
     shortfallBalance: { value: scale6(outcome.reserve.shortfall) },
   });
+};
+
+test.serial(
+  'liquidate ATOM',
+  checkFlow1,
+  { collateralBrandKey: 'ATOM', managerIndex: 0 },
+  {},
+);
+
+test.serial('add STARS collateral', async t => {
+  const { controller, buildProposal } = t.context;
+
+  t.log('building proposal');
+  const proposal = await buildProposal({
+    package: 'inter-protocol',
+    packageScriptName: 'build:add-STARS-proposal',
+  });
+
+  for await (const bundle of proposal.bundles) {
+    await controller.validateAndInstallBundle(bundle);
+  }
+  t.log('installed', proposal.bundles.length, 'bundles');
+
+  t.log('launching proposal');
+  const bridgeMessage = {
+    type: 'CORE_EVAL',
+    evals: proposal.evals,
+  };
+  t.log({ bridgeMessage });
+
+  const { EV } = t.context.runUtils;
+  /** @type {ERef<import('../../src/types.js').BridgeHandler>} */
+  const coreEvalBridgeHandler = await EV.vat('bootstrap').consumeItem(
+    'coreEvalBridgeHandler',
+  );
+  await EV(coreEvalBridgeHandler).fromBridge(bridgeMessage);
+
+  t.context.refreshAgoricNamesRemotes();
+
+  t.log('restart-vats proposal executed');
+  t.pass(); // reached here without throws
 });
+
+test.serial(
+  'liquidate STARS',
+  checkFlow1,
+  { collateralBrandKey: 'STARS', managerIndex: 1 },
+  {},
+);